fix: more eslint magic

+++ src/functions/check4Lockdowns.js
@@ -1,30 +1,32 @@
-const DB = require("../structures/schemas/lockdown.js");
+const DB = require('../structures/schemas/lockdown.js');
 
 function check4Lockdowns(client) {
-  DB.find().then((data) => {
-    data.forEach(async (data) => {
-      const channel = client.guilds.cache
-        .get(data.guildId)
-        .channels.cache.get(data.channelId);
-      if (!channel) return;
+	DB.find().then((data) => {
+		data.forEach(async () => {
+			if (!data) return;
 
-      if (data.timeLocked < Date.now()) {
-        await DB.deleteOne({ channelId: channel.id });
+			const channel = client.guilds.cache
+				.get(data.guildId)
+				.channels.cache.get(data.channelId);
+			if (!channel) return;
 
-        return channel.permissionOverwrites.edit(data.guildId, {
-          SendMessages: null,
-        });
-      }
+			if (data.timeLocked < Date.now()) {
+				await DB.deleteOne({ channelId: channel.id });
 
-      setTimeout(async () => {
-        await DB.deleteOne({ ChannelID: channel.id });
+				return channel.permissionOverwrites.edit(data.guildId, {
+					SendMessages: null,
+				});
+			}
 
-        return channel.permissionOverwrites.edit(data.guildId, {
-          SendMessages: null,
-        });
-      }, data.timeLocked - Date.now());
-    });
-  });
+			setTimeout(async () => {
+				await DB.deleteOne({ ChannelID: channel.id });
+
+				return channel.permissionOverwrites.edit(data.guildId, {
+					SendMessages: null,
+				});
+			}, data.timeLocked - Date.now());
+		});
+	});
 }
 
 module.exports = { check4Lockdowns };
